fix: whiteboard 분리
