+++ client/app/(whiteboard)/_component/_toolbar-component/stickers.tsx
