--- client/app/(workplace)/(text-editor)/(routes)/text-editor.tsx
@@ -1,5 +1,0 @@
-const TextEditor = () => {
-  return <div></div>;
-};
-
-export default TextEditor;
