--- client/app/(workplace)/layout.tsx
@@ -1,14 +1,0 @@
-interface LayoutProps {
-  children: React.ReactNode;
-}
-
-const Layout = ({ children }: LayoutProps) => {
-  return (
-    <div>
-      {/* <Header/> */}
-      {children}
-    </div>
-  );
-};
-
-export default Layout;
